verilog_flow/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "verilog-flow"
version = "3.0.0"
description = "VeriFlow-Agent 3.0: Industrial-grade Verilog code generation with timing awareness"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "VeriFlow Team"}
]
keywords = ["verilog", "rtl", "fpga", "asic", "code-generation", "timing-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0.0",
    "jinja2>=3.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-parser>=2.0.0",
]
eda = [
    "cocotb>=0.7.0",
    "cocotb-bus>=0.2.0",
    "find-libpython>=0.3.0",
]

[project.scripts]
verilog-flow = "verilog_flow.cli.main:main"
vf-codegen = "verilog_flow.cli.codegen:main"
vf-sim = "verilog_flow.cli.simulate:main"
vf-synth = "verilog_flow.cli.synthesize:main"

[project.urls]
Homepage = "https://github.com/veriflow/verilog-flow"
Documentation = "https://verilog-flow.readthedocs.io"
Repository = "https://github.com/veriflow/verilog-flow"
Issues = "https://github.com/veriflow/verilog-flow/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["verilog_flow*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"verilog_flow" = [
    "schemas/*.json",
    "templates/*.j2",
    "templates/*.v",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["verilog_flow"]
branch = true
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
precision = 2
show_missing = true
skip_covered = false
